chore: dele opp søknadoversetter i domener
